Add script to view the github pages locally
We need a development webserver to view the calculator locally,
because it does AJAX requests and those don't work over the file:// protocol.

This repo is just a collection of static files,
so there was no platform yet that could act as a webserver.
I've decided to go for ruby because that is what github pages uses
if you want more than just static files.
https://help.github.com/articles/using-jekyll-as-a-static-site-generator-with-github-pages/